qcacmn: Validate the input puncture pattern
The default puncturing input set for any channel is
ALL_SCHANS_PUNC (0xFFFFFFFF) and regulatory considers this as a
valid puncturing pattern. Hence the SP PSD power is set as 0
and the best power mode chosen becomes LPI though AFC is completed
on all SP channels.
Hence validate the input puncture pattern in the regulatory and if
the puncturing pattern is invalid, do not compute SP PSD power values
for the given puncture pattern.

Change-Id: Ib4366eb2de5a0440c649f7beb91f4eca097e82e6
CRs-Fixed: 3331572