qcacld-3.0: Add pwrmode argument to use super channel list
The current channel list alone is not enough to represent the
capability of the chip or device. Given a channel, in many
cases it may be required to know all the power modes that are
supported by this channel.

Additionally add the arguments of power mode and puncture bit map
in the caller API wlan_reg_decide_6ghz_power_within_bw_for_freq.

Change-Id: Ib96c1a15d4696ddb36cc27853a2ff0d4e0cb5cdc
CRs-Fixed: 3359248