qcacld-3.0: Skip center frequency check for 320 MHz BW
In all Wi-Fi 7 STA SW implementations, ignore the CCFS0
field in the EHT operation element when the channel BW
is 160 or 320 MHz as per spec, thus CCFS0 is set to an
invalid value e.g. 0 in EHT operation element. Hence,
skip center frequency check when bandwidth is 160 MHz
or 320 MHz and avoid dropping of bandwidth to 20 MHz.

Change-Id: Ic2fe8f74d8372f988b4fed8d926b5186678ce9d1
CRs-Fixed: 3338639