qca-wifi-oss: Consider BW for finding valid channel
When user configures channel "auto" and mode EHT320,
driver calculates the first valid frequency for the
AP to come up. The current algorithm only checks if
the channel is not disabled and not in NOL. For example,
the first available frequency which is not disabled
and not in NOL is 5180MHz. Driver selects channel 36
as valid frequency and proceeds further for checking
if the phymode is allowed. The maximum bandwidth
supported by channel 36 is EHT160 and user configured
BW is EHT320. The phymode gets rejected and NULL channel
is returned. This results in VAP not coming up.

Solution:
Consider the maximum supported BW of the channel during
the calculation of first valid channel.

Change-Id: I404148a8d2666099563db5a6b2ccc1c2e791dc8d
CRs-Fixed: 3365748