qca-wifi: Set half/quarter rate flags in multiple_vdev_restart
When Independent AP Repeater VAP is enabled, ACS brings up
Independent RE's AP VAP in a channel (say 4950). The STA VAP
identifies ROOT AP in a different channel (4945 Half Rate)
and tries to set the  Root AP channel in RE using multiple_vdev_restart.
In the current implementation, multiple_vdev_restart does not
set the Half/Quarter channel flag before sending the WMI command to the FW.
Which means the channel is treated as a FULL/20MHz channel by the FW.
However, since the root AP is in a Half rate channel,
the STA cannot connect to the root AP

To Fix this problem,
set the half/quarter rate flags in multiple_vdev_restart.

Change-Id: I2d25fd3f500d335894680a49b53324d70f2dad85
CRs-Fixed: 3360238