qcacld-3.0: In get_channel check link VDEV's assoc state
Kernel calls get_channel callback registered in cfg80211_ops to
fetch the channel information.
For a two or more links MLO connection, the state of assoc VDEV
will be set to associated before get_channel API is called, but
for partner link VDEV might still be in unassociated state and
get_channel request on parter link will lead to sending invalid
channel information to kernel and result in kernel warning.

If partner VDEV is not in associated state, return error to avoid
kernel warning.

Change-Id: I7ac3f6e08301419a9dc916e59f611537ccf88d45
CRs-Fixed: 3392518