qcacld-3.0: Ignore CSA request for invalid channel
In present scenario, STA disconnects with AP if it receives
invalid channel in CSA IE. In this case STA shouldn't
disconnect with AP as this request may come from a spoof AP.

Ignore this CSA request as it might be from spoof AP and
if it is from genuine AP heart beat failure happens and
results in disconnection. After disconnection DUT may
reconnect to same or other APs.

Change-Id: I554bda1c3a0aa96e97d7fdf4dc160be8a2e7f452
CRs-Fixed: 3390251