qcacld-3.0: Define WLAN_HDD_DEFLINK_IDX and active links
Define WLAN_HDD_DEFLINK_IDX macro which points to default
index in link info array.

Introduce a bitmap to hold the status of links in the array
which are active/opened. This bitmap will be used to check
whether the link info at a particular array index is active
or not.

Change-Id: Ibe9f6694e74ea0f566c586444f4e17d1fd433e66
CRs-Fixed: 3448606