qca-wifi-oss: Validate bangradar params on agile channel
A FW assert is when radar is simulated on the agile channel using
bangradar command. Example: "radartool -i wifix bangradar 0 0 10 1".

If the agile feature is disabled, the HOST does not program an agile
channel. However, in this situation, if the HOST tries to inject radar,
the FW crashes as no agile channel was programmed in the FW.

Do not send agile radar params to FW when adfs feature is not enabled.

Change-Id: Idad0ef9458e0e54b6805a32dadc491031bde5406
CRs-Fixed: 3413455