qcacld-3.0: Enlarge the array size of qdf_nbuf_history
Given the fact that the RX nbuf history array size is not enough,
sometimes the dump fails to capture the problematic buffer address
records, leading to failing to identify sources of buffers. This
change is to increase the array size of qdf_nbuf_history, which
makes it possible to record more RX history buffers to be traced.

Change-Id: I7982873daf6fafcb31e77d2aec380f1c00401ee5
CRs-Fixed: 3405585