qcacld-3.0: Wrapper API to check duplicate session
Introduce new API which checks for duplicate session on all the
valid addresses in the adapter.

The existing sme_check_for_duplicate_session() API is extended
to take list of addresses to check for duplicate entry.
The list argument is expected to be NULL terminated to exit the
loop.

Change-Id: I71629aa89e901041d787de18e2305dbbbca5aa50
CRs-Fixed: 3524015