qcacld-3.0: Remove dp_protect_entry_count atomic variable
The following patch removed the use of hdd_wait_for_dp_tx():
    qcacld-3.0: Update HDD to use DP component APIs
    Change-Id I377aecc4343e75dd17d279a2eb84fc49e737e784

Since the function is obsolete, remove it.

Note that this removes the only code that depends upon the
dp_protect_entry_count atomic variable, which means that is also
obsolete, as are the hdd_dp_ssr_protect() and hdd_dp_ssr_unprotect()
functions that increment and decrement it.

Change-Id: I02621dabf2e67a39259c4a51a604312f1927f72d
CRs-Fixed: 3424826