qcacld-3.0: Remove obsolete tAniSirLim counters
The tAniSirLim struct contains multiple counters. Although all of them
are initialized, only some of them are ever incremented, and very few
of them actually have their values read. Since it is pointless to keep
counters that are never examined, remove the ones that are never read.

As part of the cleanup replace the tAniSirLim typedef with a properly
named struct to align with the coding standard.

Change-Id: I8b306ff9ae92a54953dd54b80c58e691612c43a0
CRs-Fixed: 3424936