qcacld-3.0: Add wake lock in cfg820211 resume for full power down
Add wake lock in cfg80211 resume for the full power down in low
power mode, and return 0 druing system resume in this case.

For the feature of Full Power Down while in low power mode, the wlan
driver reinit happened asynchronously during system resume, so the
cfg80211 resume happened before wlan driver reinit, so it will return
error, it is not correct action for this feature in this case, so
changed to return 0 in this case.

As the the wlan driver reinit happened asynchronously during system
resume, and it is triggered by the fw ready indicate msg to mhi, it
may miss if it happened before system resume, so adding the wake lock
with timeout 5 seconds to make the system keep awake to make mhi
receive wlan fw ready indicate msg successfully.

Change-Id: Ieaa9c91ea6e6a5a92754afd9fd0e642e4d7f7be8
CRs-Fixed: 3426838