qcacld-3.0: Fix reconnect to some APs fail issue on SAP-STA mode
This issue is caused by wlan driver dropped the EAPOL-Key 1 which
sent from some APs and make our STA connect to some APs fail.

The reason why wlan driver dropped the EAPOL-Key 1 is that wlan driver
process the EAPOL-Key 1 packet before association completion handler
called. This case only happened the SAP-STA mode and IPA enabled and
there is external connected to our SAP, and casued the association
completion handler later than processing the EAPOL-Key 1 packet.

This is a compatibility issue and only happened on some APs,
because some APs will not re-send the EAPOL-Key 1 after a few
while if it did not receive the peer STA's EAPOL-Key 2 packet.

For the other phones, this issue will not happen becasue we can
see they re-send the EAPOL-Key 1 packet if the EAPOL packet
is dropped by our wlan driver in this case.

Change-Id: Ibe19e12e2d6923cbf8274e302981207213b8d622
CRs-Fixed: 3416606