qcacld-3.0: Fix to non-pre-cac adapter cleanup
If the non-pre-cac adapter stop got trigger first, non-pre-cac adapter will
get close but the pre-cac vdev is not cleaned up.

So, don't need to check for is pre-cac ON for non-pre-cac adapter.

Remove check for is pre-cac ON for non-pre-cac adapter in
hdd_stop_adapter_ext and wait for the pre-cac cleanup to complete first.

Change-Id: I03b46b7db849f45f7625bdd0666f63ba7820e0f1
CRs-Fixed: 3431907